requirements.txt
# Core dependencies for DataHub workflows and scripts
pyyaml==6.0.1
requests==2.31.0
python-dotenv==1.0.0
jsonschema==4.19.0
pytest==7.4.2
black==23.9.1
acryl-datahub